Add files via upload
This is the Java GUI Application for creating a Crossword Puzzle. Currently, it can accept both clues and answers and validate them and display the entries in a list of the applications. Functionality to save the puzzle has not been added yet. This GUI is developed by using the built-in Swing components of Java supported by OpenJDK version 11.x.