Add files via upload
This is the Java Swing source code file for the FlashCards program. 
This program has been compiled and tested by using OpenJDK 11.x in both Windows 10 and Linux Mint 21.x Operating Systems.
This program accepts Questions and Answers for FlashCards from user through the GUI and allows the user to save them in deck files in text format (as .txt files) in the same directory of the application.